Added grpc call to send new transaction to pbft nodes
